requirements.txt
flask
google-cloud-secret-manager
requests